extract helper function out
