+++ backend/presentation/Api/upload_tex_command.py
@@ -4,7 +4,7 @@ from rest_framework.response import Response
 from presentation.models import Level, RootLevel
 from presentation.Serializers.rootLevel_serializer import RootLevelSerializer
 from presentation.Serializers.level_serializers import LevelSerializer
-from presentation.Api.utils import parseTexCommandFromFile
+import re
 '''
 Api request:
 		PUT http://localhost:8000/uploadTex/[book id]/
@@ -12,6 +12,26 @@ Api request:
 		Headers:
 			Content-Disposition:   form-data;filename="xxx.tex"
 '''
+
+
+def parseTexCommandFromFile(file_content):
+    new_command = r"(\\newcommand\{\\.*\})(\[[0-9]?\])+(.*)(\%.*)"
+    pattern = re.compile(new_command)
+    matched_commands = []
+    tex = file_content.splitlines()
+    for line in tex:
+        matchObj = re.search(pattern, line)
+        if matchObj:
+            command = matchObj.group(0)
+            tex = re.compile(r"(\\newcommand\{\\.*\})(\[[0-9]?\])+\{(.*)\}")
+            comment = re.compile(r"\%.*")
+            matchTex = re.search(tex, command).group(0)
+            matchComment = re.search(comment, command).group(0)
+            matched_commands.append({
+                "tex": matchTex,
+                "note": matchComment[1:]
+            })
+    return matched_commands
 
 
 class UploadTexCommand(views.APIView):
